:truck::art: split onboarding service into multi modules
