+++ app/tests/admin/test_onboarding.py
@@ -13,9 +13,8 @@ from mockito import verify, when
 
 from app.event_handling.sse_listener import SseListener
 from app.exceptions.cloud_api_error import CloudApiException
-from app.services import onboarding
+from app.services import acapy_ledger, acapy_wallet, onboarding
 from app.services.acapy_wallet import Did
-from app.services.onboarding import acapy_ledger, acapy_wallet
 from app.tests.util.mock import to_async
 from shared.util.mock_agent_controller import get_mock_agent_controller
